+++ mds/apis/agency_api/v0_x/service_areas.py
@@ -7,7 +7,6 @@ from mds import enums
 from mds import models
 from mds.access_control.permissions import require_scopes
 from mds.access_control.scopes import SCOPE_AGENCY_API
-from mds.apis import utils
 
 
 class MultiPolygonField(serializers.Field):
@@ -28,15 +27,6 @@ class AreaSerializer(serializers.ModelSerializer):
     service_area_id = serializers.UUIDField(
         source="id", help_text="Unique Area identifier (UUID)"
     )
-    start_date = utils.UnixTimestampMilliseconds(
-        source="creation_date",
-        help_text="Date at which this service area became effective",
-    )
-    end_date = utils.UnixTimestampMilliseconds(
-        source="deletion_date",
-        required=False,
-        help_text="If exists, Date at which this service area was replaced.",
-    )
     area = MultiPolygonField(source="polygons.all")
     type = serializers.ChoiceField(
         source="area_type",
@@ -46,7 +36,7 @@ class AreaSerializer(serializers.ModelSerializer):
 
     class Meta:
         model = models.Area
-        fields = ("service_area_id", "start_date", "end_date", "area", "type")
+        fields = ("service_area_id", "area", "type")
 
 
 class AreaViewSet(viewsets.ReadOnlyModelViewSet):