+++ mds/factories.py
@@ -583,8 +583,6 @@ class Area(factory.DjangoModelFactory):
     class Meta:
         model = models.Area
 
-    creation_date = datetime.datetime(2018, 8, 1, tzinfo=pytz.utc)
-    deletion_date = None
     label = ""
 
     @factory.post_generation
@@ -616,8 +614,6 @@ class Polygon(factory.DjangoModelFactory):
     class Meta:
         model = models.Polygon
 
-    creation_date = datetime.datetime(2018, 8, 1, tzinfo=pytz.utc)
-    deletion_date = None
     label = ""
     geom = district10
     properties = {}