+++ mds/models.py
@@ -1,7 +1,6 @@
 """
 Database description
 """
-import datetime
 import json
 import uuid
 
@@ -221,8 +220,6 @@ class EventRecord(models.Model):
 
 class Polygon(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4)
-    creation_date = models.DateTimeField(default=utils.timezone.now)
-    deletion_date = models.DateTimeField(null=True, blank=True)
     label = UnboundedCharField(default="", blank=True, db_index=True)
     geom = gis_models.PolygonField()
     properties = pg_fields.JSONField(default=dict, encoder=encoders.JSONEncoder)
@@ -233,10 +230,6 @@ class Polygon(models.Model):
 
 class Area(models.Model):
     id = models.UUIDField(primary_key=True, default=uuid.uuid4)
-    creation_date = models.DateTimeField(
-        default=datetime.datetime(2012, 1, 1, tzinfo=datetime.timezone.utc)
-    )
-    deletion_date = models.DateTimeField(null=True, blank=True)
     label = UnboundedCharField(default="", blank=True, db_index=True)
     polygons = models.ManyToManyField(Polygon, blank=True, related_name="areas")
     providers = models.ManyToManyField(Provider, blank=True, related_name="areas")