+++ tests/apis/agency_api/v0_x/test_areas.py
@@ -1,4 +1,3 @@
-import datetime
 import pytest
 
 from mds import factories
@@ -20,11 +19,7 @@ def test_areas_metadata(client):
 @pytest.mark.django_db
 def test_areas_detail(client, django_assert_num_queries):
     provider = factories.Provider(name="Test provider")
-    area = factories.Area(
-        creation_date=datetime.datetime(2016, 1, 1, tzinfo=datetime.timezone.utc),
-        deletion_date=datetime.datetime(2017, 1, 1, tzinfo=datetime.timezone.utc),
-        providers=[provider],
-    )
+    area = factories.Area(providers=[provider])
     other_provider = factories.Provider(name="Test other provider")
 
     response = client.get("/mds/v0.x/service_areas/%s/" % area.pk)
@@ -53,8 +48,6 @@ def test_areas_detail(client, django_assert_num_queries):
     assert response.status_code == 200
     assert response.data == {
         "service_area_id": str(area.pk),
-        "start_date": 1_451_606_400_000,
-        "end_date": 1_483_228_800_000,
         "area": {
             "coordinates": [
                 [[[0.0, 0.0], [0.0, 50.0], [50.0, 50.0], [50.0, 0.0], [0.0, 0.0]]]