+++ CHANGELOG.rst
@@ -7,7 +7,7 @@ Changelog
 - Add token authentication by specifying token in browser url
 - Fix compiled translations not embedding in releases
 - Events pushed by providers now take precedence over pulled ones.
-
+- Delete creation_date and deletion_date fields on Area and Polygon
 
 0.5.12 (2019-04-05)
 -------------------